requirements.txt
requests-html==0.9.0
html2text==2018.1.9
zope.cachedescriptors==4.3.1
numpy==1.16.4
scikit-learn==0.21.1
sklearn
scipy==1.3.0
nltk==3.4.3
